% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luccpl_format.R
\docType{data}
\name{and_or_bool}
\alias{and_or_bool}
\title{and_or_bool}
\usage{
and_or_bool(connector_list, n_relations)
}
\arguments{
\item{connector_list}{is a list of and/or to convert in 0/1}

\item{n_relations}{number of relations to the query}
}
\value{
function to return 0 to 'and' and 1 to 'or'
}
\description{
This function convert the conjunction operator and/or in 0/1 to make query array.
}
\author{
Carlos Alexandre Romani
}
\keyword{datasets}
