% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{create_brick}
\alias{create_brick}
\title{Brick creation}
\usage{
create_brick(path)
}
\arguments{
\item{path}{where is the classified images in .tif format}
}
\value{
Brick of spatiotemporal data
}
\description{
Return a brick of spatiotemporal data from classified images
}
\author{
Carlos Alexandre Romani
}
