% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{event}
\alias{event}
\title{Lucc event}
\usage{
(rbrick, query_array)
}
\arguments{
\item{rbrick}{input brick}

\item{query_array}{result of parse expression}
}
\value{
Brick of result (bool)
}
\description{
Analysis
}
\author{
Carlos Alexandre Romani
}
