% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{export_brick}
\alias{export_brick}
\title{Export Brick}
\usage{
export_brick(raster, filename)
}
\arguments{
\item{path}{where is the classified images in .tif format}
}
\description{
Write a brick in a file
}
\author{
Carlos Alexandre Romani
}
