% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luccpl_format.R
\docType{data}
\name{format1}
\alias{format1}
\title{format1}
\usage{
format1(x)
}
\arguments{
\item{x}{patterns}
}
\value{
formated patterns
}
\description{
Function to format many patterns to parse. internally used
}
\author{
Carlos Alexandre Romani
}
\keyword{datasets}
