% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_brick}
\alias{import_brick}
\title{Import Brick}
\usage{
import_brick(filename)
}
\arguments{
\item{filename}{of a exported brick}
}
\value{
Brick of spatiotemporal data
}
\description{
Import a brick of spatiotemporal data
}
\author{
Carlos Alexandre Romani
}
