% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luccpl_plot_graph.R
\docType{data}
\name{luccpl_barplot_data}
\alias{luccpl_barplot_data}
\title{Plot input data in bar or line graph}
\usage{
luccpl_barplot_data(df=NULL, dates=NULL, graph_title=NULL, style='bar', colors=NULL, path_save_jpeg=NULL)
}
\arguments{
\item{df}{Dataframe generated from the LuccPL::luccpl_count(for_time_step = TRUE), when luccpl_count is applied to the input rasterBrick.}

\item{dates}{Integer. A vector with the timeline dates.}

\item{style}{Character. 'bar' or 'line' graph.}

\item{graph_title}{Character. The name of graph.}

\item{colors}{Character. A vector with color names of original raster, same order.}

\item{path_save_jpeg}{Character. The path of output graph in .jpeg format.}
}
\value{
A jpeg file with bar or line graph that represents the input data.
}
\description{
Creates a line or bar graph that represents the input data.
}
\author{
Carlos Alexandre Romani
}
