% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luccpl_count.R
\docType{data}
\name{luccpl_count}
\alias{luccpl_count}
\title{Count}
\usage{
luccpl_count(rbrick, for_time_step = FALSE, metadata = NULL, dates = NULL)
}
\arguments{
\item{rbrick}{S4. A spatiotemporal raster brick}

\item{for_time_step}{Boolean. FALSE returns a single raster with sum of timeserie values (use for query result only). 
TRUE returns a table with histogram for each land use.}

\item{metadata}{Character. A vector with the land use patterns of the input data}

\item{dates}{Integer. A vector that represents the timeline}
}
\value{
- for_time_step = FALSE, returns a single raster, the value of each cell being 
   the result of the sum of true values over time (only applicable to the query result raster_brick)
 - for_time_step = TRUE, returns a table with the number of cells of each class and in each instant of time.
   If applied to the rasterbrick result of the query, it returns a table with the true and false for each instant of time.
}
\description{
Queries performed with the event function return a three-dimensional matrix (x, y, t) the count
 function performs a vector summation in time or space of these results, allowing for more tangible analyzes. 
 The result can be a two-dimensional matrix (x, y) with the true count for each location or a table with the 
 true count for each year. If applied in the classifications rasterbrick, it returns a table with the number 
 of cells of each class and in each time instant.
}
\author{
Carlos Alexandre Romani
}
\keyword{datasets}
