% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luccpl_relations.R
\docType{data}
\name{luccpl_during}
\alias{luccpl_during}
\title{Relation During}
\usage{
luccpl_during(pattern_list1 = NULL, pattern_list2 = NULL, date1 = NULL, date2 = NULL, dates = NULL, metadata = NULL)
}
\arguments{
\item{pattern_list1}{Character. List of patterns used for the query.}

\item{pattern_list2}{Not used in this function, as it does not relate different patterns.}

\item{date1}{Integer. Initial date}

\item{date2}{Integer. Final date}

\item{dates}{Integer. A vector that represents the timeline}

\item{metadata}{Character. A vector with the land use patterns of the input data}
}
\value{
A query_array that can be grouped with others to perform the query
}
\description{
Classifies as true all the locations and instants in which
one of the classes present in p is contained in the instants of the interval ]ti, tj[
}
\author{
Carlos Alexandre Romani
}
\keyword{datasets}
