% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luccpl_event.R
\docType{data}
\name{luccpl_event}
\alias{luccpl_event}
\title{Event}
\usage{
luccpl_event(rbrick, query_array)
}
\arguments{
\item{rbrick}{S4. A spatiotemporal raster brick}

\item{query_array}{Integer. Query matrix generated by the luccpl_query function,
containing the information to perform the queries in numerical form.}
}
\value{
A Boolean data cube, with answers to the questions asked.
}
\description{
Main function of the package, where the processing is performed based on the query subfunctions.
This function takes a cube of classified land use data as input, together with a query array.
The query array is applied to each location (x, y) in the data cube, performing analyzes, and returning
as a result a Boolean data cube, with answers to the questions asked.
}
\author{
Carlos Alexandre Romani
}
\keyword{datasets}
