% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luccpl_relations.R
\docType{data}
\name{luccpl_evolve}
\alias{luccpl_evolve}
\title{Relation Evolve}
\usage{
luccpl_evolve(pattern_list1 = NULL, pattern_list2 = NULL, date1 = NULL, date2 = NULL, dates = NULL, metadata = NULL)
}
\arguments{
\item{pattern_list1}{Character. First list of patterns used for the query.}

\item{pattern_list2}{Character. Second list of patterns used for the query}

\item{date1}{Integer. Initial date}

\item{date2}{Integer. Final date}

\item{dates}{Integer. A vector that represents the timeline}

\item{metadata}{Character. A vector with the land use patterns of the input data}
}
\value{
A query_array that can be grouped with others to perform the query
}
\description{
Classifies as true all the locations and times when one of the classes in pi evolved
into one of the classes in pj within the range [ti, tj].
}
\author{
Carlos Alexandre Romani
}
\keyword{datasets}
