% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luccpl_relations.R
\docType{data}
\name{luccpl_metby}
\alias{luccpl_metby}
\title{Relation Met By}
\usage{
luccpl_metby(pattern_list1 = NULL, pattern_list2 = NULL, date1 = NULL, date2 = NULL, dates = NULL, metadata = NULL)
}
\arguments{
\item{pattern_list1}{Character. List of patterns used for the query.}

\item{pattern_list2}{Not used in this function, as it does not relate different patterns.}

\item{date1}{Integer. Date referring to pattetn_list1}

\item{date2}{Not used in this function because pattern_list2 is not used}

\item{dates}{Integer. A vector that represents the timeline}

\item{metadata}{Character. A vector with the land use patterns of the input data}
}
\value{
A query_array that can be grouped with others to perform the query
}
\description{
Classifies as true all the locations and moments in which
one of the classes in p is present at the same time and after time ]t
}
\author{
Carlos Alexandre Romani
}
\keyword{datasets}
