% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luccpl_plot.R
\docType{data}
\name{luccpl_plot_input}
\alias{luccpl_plot_input}
\title{Plot Input Data}
\usage{
luccpl_plot_input(X,date,pattern,color,pathMPinput,map_title,Width,Height)
}
\arguments{
\item{X}{S4. A spatiotemporal raster brick}

\item{date}{Integer. A vector that represents the timeline}

\item{pattern}{Character. A vector with the land use patterns of the input data}

\item{color}{Character. A vector with the colors to represent land use}

\item{pathMPinput}{Character. Path to save .jpeg of the raster brick}

\item{map_title}{Character. Title of multiplot}

\item{Width}{Integer. Dimension of multiplot}

\item{Height}{Integer. Dimension of multiplot}
}
\value{
Multiplot of input raster brick
}
\description{
This function creates an image in .jpeg format with all layers of the raster brick, 
as well as caption of the land use classes
}
\author{
Carlos Alexandre Romani
}
\keyword{datasets}
