% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luccpl_plot.R
\docType{data}
\name{luccpl_plot_output}
\alias{luccpl_plot_output}
\title{Plot Output Data}
\usage{
luccpl_plot_output(X,date,pathMPinput,map_title,Width,Height)
}
\arguments{
\item{X}{S4. A spatiotemporal raster brick, result of query}

\item{date}{Integer. A vector that represents the timeline}

\item{map_title}{Character. Title of multiplot}

\item{Width}{Integer. Dimension of multiplot}

\item{Height}{Integer. Dimension of multiplot}

\item{pathMPinput}{Character. Path to save .jpeg of the raster brick}
}
\value{
Multiplot of output raster brick
}
\description{
This function creates an image in .jpeg format with all layers of the output raster brick
}
\author{
Carlos Alexandre Romani
}
\keyword{datasets}
