% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luccpl_format.R
\docType{data}
\name{luccpl_query}
\alias{luccpl_query}
\title{Query}
\usage{
luccpl_query(FUN_list)
}
\arguments{
\item{FUN_list}{List of functions (relations)}
}
\value{
A query_array to perform the query on luccpl_event function.
}
\description{
Here, the query queries generated by the different query functions are organized and joined,
 the result being a query_array
}
\author{
Carlos Alexandre Romani
}
\keyword{datasets}
