% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luccpl_true_single_raster.R
\docType{data}
\name{luccpl_true}
\alias{luccpl_true}
\title{True}
\usage{
luccpl_true(rbrick)
}
\arguments{
\item{rbrick}{S4. A spatiotemporal raster brick, output from luccpl_event.}
}
\value{
A single layer raster with true in locations with true for locations with at least one true on the timeline
}
\description{

}
\author{
Carlos Alexandre Romani
}
\keyword{datasets}
