% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luccpl_variance.R
\docType{data}
\name{luccpl_variance}
\alias{luccpl_variance}
\title{Variance}
\usage{
luccpl_variance(rbrick)
}
\arguments{
\item{rbrick}{S4. A spatiotemporal raster brick}
}
\value{
a single raster, the value of variance to each cell
}
\description{
With a similar purpose to the count function, this, when applied to a three-dimensional matrix (x, y, t), 
generates a two-dimensional matrix (x, y) with the number of variations that occurred over time for each cell.
}
\author{
Carlos Alexandre Romani
}
\keyword{datasets}
