% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luccpl_format.R
\docType{data}
\name{mdata}
\alias{mdata}
\title{Mdata}
\usage{
mdata(pattern_list, metadata)
}
\arguments{
\item{pattern_list}{list of patterns used in query}

\item{metadata}{of land use classes for input data}
}
\value{
a list of digital numbers of classifications.
}
\description{
This function make relation between metadata and digital number in data input.
}
\author{
Carlos Alexandre Romani
}
\keyword{datasets}
