% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_input}
\alias{plot_input}
\title{Plot input}
\usage{
plot_input(X, date, pattern, color)
}
\arguments{
\item{X}{is a brick}

\item{date}{is a vector with a time serie of dates}

\item{pattern}{is a metadata of land use classes}

\item{color}{is the color scheme to represent each classe}
}
\description{
Create a multiplot from input classified images
}
\author{
Carlos Alexandre Romani
}
