% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_output}
\alias{plot_output}
\title{Plot output}
\usage{
plot_output(X, date)
}
\arguments{
\item{X}{is a brick}

\item{date}{is a vector with a time serie of dates}
}
\description{
Create a multiplot of output
}
\author{
Carlos Alexandre Romani
}
