% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luccpl_format.R
\docType{data}
\name{tdata}
\alias{tdata}
\title{Tdata}
\usage{
tdata(date, dates)
}
\arguments{
\item{date}{date from query}

\item{dates}{dates from timeserie}
}
\value{
function to return the time-step for each data.
}
\description{
This function make relation between dates and timestep.
}
\author{
Carlos Alexandre Romani
}
\keyword{datasets}
